+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 4.7.0 - 23/12/25
 * NOVO: Campo de CPF/CNPJ
 * NOVO: Campo de Bairro.
+* Ajuste: barra de frete grátis.
 
 # 4.6.0 - 15/12/25
 * NOVO: Atualização na logica de frete grátis.